# Improve the performance of Go generated marshaling/encoding code