--- a/lib/xdrgen/generators/go.rb
+++ b/lib/xdrgen/generators/go.rb
@@ -174,7 +174,11 @@ module Xdrgen
           render_binary_interface out, name(defn)
         when AST::Definitions::Typedef ;
           render_typedef out, defn
-          render_binary_interface out, name(defn)
+          # TODO: Support defining binary interface on typedefs with optional
+          # types. https://github.com/stellar/xdrgen/issues/61
+          if defn.sub_type != :optional
+            render_binary_interface_typedef out, defn
+          end
         when AST::Definitions::Const ;
           render_const out, defn
         end