--- a/lib/xdrgen/generators/go.rb
+++ b/lib/xdrgen/generators/go.rb
@@ -1,5 +1,3 @@
-require 'pry'
-
 module Xdrgen
   module Generators