@leighmcculloch 2021-09-24
I'm reasonably confident that this change is accurate and complete. I used the Stellar XDR and tests that live in the `github.com/stellar/go` `xdr` and `txnbuild` packages, along with some manual tests of some edge cases like `SponsorshipDescriptors`, to check for consistency. If anyone has ideas for other ways to test this, please let me know.