@ire-and-curses 2021-10-05
> LGTM! It should improve ingestion speed in Horizon. Probably fixes [stellar/go#2689](https://github.com/stellar/go/issues/2689) and [stellar/go#3256](https://github.com/stellar/go/issues/3256).

Note that this PR doesn't improve unmarshaling, so there may be more we can squeeze out in the future.