@bartekn 2021-10-05
Right, but `TransactionProcessor` is only marshaling.