@leighmcculloch 2021-10-05
> `TransactionProcessor` is only marshaling

I'm wondering if this is an opportunity to test this new marshaling code. How would we test that the marshaling is consistent in Horizon with this change? If Horizon marshals XDR during ingestion, if we rebuilt Horizon using xdr generated by this PR, and ran a full history ingestion, would that be sufficient at identifying any problems marshaling that data? Would Horizon error if the marshaling resulted in different XDR?