@bartekn 2021-10-05
We can easily confirm this by [`verify-range`](https://github.com/stellar/go/tree/master/services/horizon/docker/verify-range) test that we run before every release (almost - we don't if there are no ingestion changes). We can run a limited (like 100k ledgers) test against a branch (like this PR branch) to confirm the result are the same.