@bartekn 2021-10-05
Also, it looks like staticcheck found some issues in generated code: https://github.com/stellar/go/pull/3957/checks?check_run_id=3705045608