@leighmcculloch 2021-10-06
@bartekn I got https://github.com/stellar/go/pull/3957 lint errors fixed. Could you kick off a verify-range test using that PR?