@bartekn 2021-10-18
@leighmcculloch I think the verify-range did not find any issues? Do you think we can merge this and https://github.com/stellar/go/pull/3957? I'm looking forward for this performance bump.