@2opremio 2021-11-08
I know this arrives very late ... but wouldn't it be better to use:

```go
type xdrEncodable interface {
	EncodeTo(e *xdr.Encoder) error
}

// Marshal writes an xdr element `v` into `w`.
func Marshal(w io.Writer, v interface{}) (int, error) {
	if encodable, ok := v.(xdrEncodable); ok {
		b := bytes.Buffer{}
		e := xdr.NewEncoder(&b)
		err := encodable.EncodeTo(e)
		return w.Write( b.Bytes())
	}
	// delegate to xdr package's Marshal
	return xdr.Marshal(w, v)
}
```

instead of

```go
type xdrType interface {
	xdrType()
}

// Marshal writes an xdr element `v` into `w`.
func Marshal(w io.Writer, v interface{}) (int, error) {
	if _, ok := v.(xdrType); ok {
		if bm, ok := v.(encoding.BinaryMarshaler); ok {
			b, err := bm.MarshalBinary()
			if err != nil {
				return 0, err
			}
			return w.Write(b)
		}
	}
	// delegate to xdr package's Marshal
	return xdr.Marshal(w, v)
}
```

Using `xdrType` is artificial an noisy