@leighmcculloch 2021-11-08
That looks cleaner.

My only hesistancy is I wouldn't want to move the bytes.Buffer up to a generic function disconnected from the type, because I hope to replace bytes.Buffer with fixed sized buffers based on max size of the types, and we can only do that if the buffer is defined inside the MarshalBinary, and if xdr.Marshal calls MarshalBinary.

We could still test on xdr.xdrEncoderable, but then call .MarshalBinary, but that doesn't completely get rid of the oddity.