@2opremio 2021-11-08
I don't think we need fixed buffers, see https://github.com/stellar/go/pull/4056